Created widget.py and refactored the code
