+++ release/scripts/mgear/rigbits/rbf_manager2/rbf_manager_ui.py
@@ -43,9 +43,9 @@ newScene callback
 Attributes:
     CTL_SUFFIX (str): suffix for anim controls
     DRIVEN_SUFFIX (str): suffix for driven group nodes
-    EXTRA_MODULE_DICT (str): name of the dict which holds additional modules
-    MGEAR_EXTRA_ENVIRON (str): environment variable to query for paths
-    TOOL_NAME (str): name of UI
+    widget.EXTRA_MODULE_DICT (str): name of the dict which holds additional modules
+    widget.MGEAR_EXTRA_ENVIRON (str): environment variable to query for paths
+    widget.TOOL_NAME (str): name of UI
     TOOL_TITLE (str): title as it appears in the ui
     __version__ (float): UI version
 
@@ -80,39 +80,12 @@ from mgear.rigbits.six import PY2
 # rbf
 from mgear.rigbits import rbf_io
 from mgear.rigbits import rbf_node
+from . import widget
 
-
-# =============================================================================
-# Constants
-# =============================================================================
-__version__ = "2.0.0"
-
-_mgear_version = mgear.getVersion()
-TOOL_NAME = "RBF Manager"
-TOOL_TITLE = "{} v{} | mGear {}".format(TOOL_NAME, __version__, _mgear_version)
-UI_NAME = "RBFManagerUI"
-WORK_SPACE_NAME = UI_NAME + "WorkspaceControl"
-
-DRIVEN_SUFFIX = rbf_node.DRIVEN_SUFFIX
-CTL_SUFFIX = rbf_node.CTL_SUFFIX
-
-MGEAR_EXTRA_ENVIRON = "MGEAR_RBF_EXTRA"
-EXTRA_MODULE_DICT = "extraFunc_dict"
-
-MIRROR_SUFFIX = "_mr"
 
 # =============================================================================
 # general functions
 # =============================================================================
-
-
-def testFunctions(*args):
-    """test function for connecting signals during debug
-
-    Args:
-        *args: Description
-    """
-    print('!!', args)
 
 
 def getPlugAttrs(nodes, attrType="keyable"):
@@ -190,16 +163,16 @@ def getEnvironModules():
     Returns:
         dict: displayName:funcObject
     """
-    extraModulePath = os.environ.get(MGEAR_EXTRA_ENVIRON, None)
+    extraModulePath = os.environ.get(widget.MGEAR_EXTRA_ENVIRON, None)
     if extraModulePath is None or not os.path.exists(extraModulePath):
         return None
-    exModule = imp.load_source(MGEAR_EXTRA_ENVIRON,
+    exModule = imp.load_source(widget.MGEAR_EXTRA_ENVIRON,
                                os.path.abspath(extraModulePath))
-    additionalFuncDict = getattr(exModule, EXTRA_MODULE_DICT, None)
+    additionalFuncDict = getattr(exModule, widget.EXTRA_MODULE_DICT, None)
     if additionalFuncDict is None:
-        mc.warning("'{}' not found in {}".format(EXTRA_MODULE_DICT,
+        mc.warning("'{}' not found in {}".format(widget.EXTRA_MODULE_DICT,
                                                  extraModulePath))
-        print("No additional menu items added to {}".format(TOOL_NAME))
+        print("No additional menu items added to {}".format(widget.TOOL_NAME))
     return additionalFuncDict
 
 
@@ -213,66 +186,6 @@ def selectNode(name):
         mc.select(name)
     else:
         print(name, "No longer exists for selection!")
-
-
-# =============================================================================
-# UI General Functions
-# =============================================================================
-
-def getControlAttrWidget(nodeAttr, label=""):
-    """Create a cmds.attrControlGrp and wrap it in a qtWidget, preserving its connection
-    to the specified attr
-
-    Args:
-        nodeAttr (str): node.attr, the target for the attrControlGrp
-        label (str, optional): name for the attr widget
-
-    Returns:
-        QtWidget.QLineEdit: A Qt widget created from attrControlGrp
-        str: The name of the created Maya attrControlGrp
-    """
-    mAttrFeild = mc.attrControlGrp(attribute=nodeAttr, label=label, po=True)
-
-    # Convert the Maya control to a Qt pointer
-    ptr = mui.MQtUtil.findControl(mAttrFeild)
-
-    # Wrap the Maya control into a Qt widget, considering Python version
-    controlWidget = QtCompat.wrapInstance(long(ptr) if PY2 else int(ptr), base=QtWidgets.QWidget)
-    controlWidget.setContentsMargins(0, 0, 0, 0)
-    controlWidget.setMinimumWidth(0)
-
-    attrEdit = [wdgt for wdgt in controlWidget.children() if type(wdgt) == QtWidgets.QLineEdit]
-    [wdgt.setParent(attrEdit[0]) for wdgt in controlWidget.children()
-     if type(wdgt) == QtCore.QObject]
-
-    attrEdit[0].setParent(None)
-    controlWidget.setParent(attrEdit[0])
-    controlWidget.setHidden(True)
-    return attrEdit[0], mAttrFeild
-
-
-def HLine():
-    """seporator line for widgets
-
-    Returns:
-        Qframe: line for seperating UI elements visually
-    """
-    seperatorLine = QtWidgets.QFrame()
-    seperatorLine.setFrameShape(QtWidgets.QFrame.HLine)
-    seperatorLine.setFrameShadow(QtWidgets.QFrame.Sunken)
-    return seperatorLine
-
-
-def VLine():
-    """seporator line for widgets
-
-    Returns:
-        Qframe: line for seperating UI elements visually
-    """
-    seperatorLine = QtWidgets.QFrame()
-    seperatorLine.setFrameShape(QtWidgets.QFrame.VLine)
-    seperatorLine.setFrameShadow(QtWidgets.QFrame.Sunken)
-    return seperatorLine
 
 
 def show(dockable=True, newSceneCallBack=True, *args):
@@ -289,8 +202,8 @@ def show(dockable=True, newSceneCallBack=True, *args):
     global RBF_UI  # Ensure we have access to the global variable
 
     # Attempt to close any existing UI with the given name
-    if mc.workspaceControl(WORK_SPACE_NAME, exists=True):
-        mc.deleteUI(WORK_SPACE_NAME)
+    if mc.workspaceControl(widget.WORK_SPACE_NAME, exists=True):
+        mc.deleteUI(widget.WORK_SPACE_NAME)
 
     # Create the UI
     RBF_UI = RBFManagerUI(newSceneCallBack=newSceneCallBack)
@@ -307,489 +220,86 @@ def show(dockable=True, newSceneCallBack=True, *args):
     return RBF_UI
 
 
-def genericWarning(parent, warningText):
-    """generic prompt warning with the provided text
-
-    Args:
-        parent (QWidget): Qwidget to be parented under
-        warningText (str): information to display to the user
-
-    Returns:
-        QtCore.Response: of what the user chose. For warnings
-    """
-    selWarning = QtWidgets.QMessageBox(parent)
-    selWarning.setText(warningText)
-    results = selWarning.exec_()
-    return results
-
-
-def promptAcceptance(parent, descriptionA, descriptionB):
-    """Warn user, asking for permission
-
-    Args:
-        parent (QWidget): to be parented under
-        descriptionA (str): info
-        descriptionB (str): further info
-
-    Returns:
-        QtCore.Response: accept, deline, reject
-    """
-    msgBox = QtWidgets.QMessageBox(parent)
-    msgBox.setText(descriptionA)
-    msgBox.setInformativeText(descriptionB)
-    msgBox.setStandardButtons(QtWidgets.QMessageBox.Ok |
-                              QtWidgets.QMessageBox.Cancel)
-    msgBox.setDefaultButton(QtWidgets.QMessageBox.Cancel)
-    decision = msgBox.exec_()
-    return decision
-
-
-class ClickableLineEdit(QtWidgets.QLineEdit):
-
-    """subclass to allow for clickable lineEdit, as a button
-
-    Attributes:
-        clicked (QtCore.Signal): emitted when clicked
-    """
-
-    clicked = QtCore.Signal(str)
-
-    def mousePressEvent(self, event):
-        if event.button() == QtCore.Qt.LeftButton:
-            self.clicked.emit(self.text())
-        else:
-            super(ClickableLineEdit, self).mousePressEvent(event)
-
-
-class TabBar(QtWidgets.QTabBar):
-    """Subclass to get a taller tab widget, for readability
-    """
-
-    def __init__(self):
-        super(TabBar, self).__init__()
-
-    def tabSizeHint(self, index):
-        width = QtWidgets.QTabBar.tabSizeHint(self, index).width()
-        return QtCore.QSize(width, 25)
-
-
-class RBFWidget(MayaQWidgetDockableMixin, QtWidgets.QMainWindow):
-
-    def __init__(self, parent=pyqt.maya_main_window()):
-        super(RBFWidget, self).__init__(parent=parent)
-
-        # UI info -------------------------------------------------------------
-        self.callBackID = None
-        self.setWindowTitle(TOOL_TITLE)
-        self.setObjectName(UI_NAME)
-        self.setAttribute(QtCore.Qt.WA_DeleteOnClose, True)
-        self.genericWidgetHight = 24
-
-    @staticmethod
-    def deleteAssociatedWidgetsMaya(widget, attrName="associatedMaya"):
-        """delete core ui items 'associated' with the provided widgets
-
-        Args:
-            widget (QWidget): Widget that has the associated attr set
-            attrName (str, optional): class attr to query
-        """
-        if hasattr(widget, attrName):
-            for t in getattr(widget, attrName):
-                try:
-                    mc.deleteUI(t, ctl=True)
-                except Exception:
-                    pass
-        else:
-            setattr(widget, attrName, [])
-
-    @staticmethod
-    def deleteAssociatedWidgets(widget, attrName="associated"):
-        """delete widget items 'associated' with the provided widgets
-
-        Args:
-            widget (QWidget): Widget that has the associated attr set
-            attrName (str, optional): class attr to query
-        """
-        if hasattr(widget, attrName):
-            for t in getattr(widget, attrName):
-                try:
-                    t.deleteLater()
-                except Exception:
-                    pass
-        else:
-            setattr(widget, attrName, [])
-
-    @staticmethod
-    def _associateRBFnodeAndWidget(tabDrivenWidget, rbfNode):
-        """associates the RBFNode with a widget for convenience when adding,
-        deleting, editing
-
-        Args:
-            tabDrivenWidget (QWidget): tab widget
-            rbfNode (RBFNode): instance to be associated
-        """
-        setattr(tabDrivenWidget, "rbfNode", rbfNode)
-
-    @staticmethod
-    def createCustomButton(label, size=(35, 27), icon=None, iconSize=None, tooltip=""):
-        stylesheet = (
-            "QPushButton:pressed { background-color: #00A6F3;}"
-        )
-        button = QtWidgets.QPushButton(label)
-        button.setMinimumSize(QtCore.QSize(*size))
-        button.setStyleSheet(stylesheet)
-        button.setToolTip(tooltip)
-        button.setIcon(pyqt.get_icon(icon, iconSize))
-        return button
-
-    @staticmethod
-    def createSetupSelector2Widget():
-        rbfVLayout = QtWidgets.QVBoxLayout()
-        rbfListWidget = QtWidgets.QListWidget()
-        rbfVLayout.addWidget(rbfListWidget)
-        return rbfVLayout, rbfListWidget
-
-    @staticmethod
-    def labelListWidget(label, attrListType, horizontal=True):
-        """create the listAttribute that users can select their driver/driven
-        attributes for the setup
-
-        Args:
-            label (str): to display above the listWidget
-            horizontal (bool, optional): should the label be above or infront
-            of the listWidget
-
-        Returns:
-            list: QLayout, QListWidget
-        """
-        if horizontal:
-            attributeLayout = QtWidgets.QHBoxLayout()
-        else:
-            attributeLayout = QtWidgets.QVBoxLayout()
-        attributeLabel = QtWidgets.QLabel(label)
-        attributeListWidget = QtWidgets.QListWidget()
-        attributeListWidget.setObjectName("{}ListWidget".format(attrListType))
-        attributeLayout.addWidget(attributeLabel)
-        attributeLayout.addWidget(attributeListWidget)
-        return attributeLayout, attributeListWidget
-
-    @staticmethod
-    def addRemoveButtonWidget(label1, label2, horizontal=True):
-        if horizontal:
-            addRemoveLayout = QtWidgets.QHBoxLayout()
-        else:
-            addRemoveLayout = QtWidgets.QVBoxLayout()
-        addAttributesButton = QtWidgets.QPushButton(label1)
-        removeAttributesButton = QtWidgets.QPushButton(label2)
-        addRemoveLayout.addWidget(addAttributesButton)
-        addRemoveLayout.addWidget(removeAttributesButton)
-        return addRemoveLayout, addAttributesButton, removeAttributesButton
-
-    def selectNodeWidget(self, label, buttonLabel="Select"):
-        """create a lout with label, lineEdit, QPushbutton for user input
-        """
-        stylesheet = (
-            "QLineEdit { background-color: #404040;"
-            "border-radius: 4px;"
-            "border-color: #505050;"
-            "border-style: solid;"
-            "border-width: 1.4px;}"
-        )
-
-        nodeLayout = QtWidgets.QHBoxLayout()
-        nodeLayout.setSpacing(4)
-
-        nodeLabel = QtWidgets.QLabel(label)
-        nodeLabel.setFixedWidth(40)
-        nodeLineEdit = ClickableLineEdit()
-        nodeLineEdit.setStyleSheet(stylesheet)
-        nodeLineEdit.setReadOnly(True)
-        nodeSelectButton = self.createCustomButton(buttonLabel)
-        nodeSelectButton.setFixedWidth(40)
-        nodeLineEdit.setFixedHeight(self.genericWidgetHight)
-        nodeSelectButton.setFixedHeight(self.genericWidgetHight)
-        nodeLayout.addWidget(nodeLabel)
-        nodeLayout.addWidget(nodeLineEdit, 1)
-        nodeLayout.addWidget(nodeSelectButton)
-        return nodeLayout, nodeLineEdit, nodeSelectButton
-
-    def createSetupSelectorWidget(self):
-        """create the top portion of the weidget, select setup + refresh
-
-        Returns:
-            list: QLayout, QCombobox, QPushButton
-        """
-        setRBFLayout = QtWidgets.QHBoxLayout()
-        rbfLabel = QtWidgets.QLabel("Select RBF Setup:")
-        rbf_cbox = QtWidgets.QComboBox()
-        rbf_refreshButton = self.createCustomButton(
-            "", (35, 25), icon="mgear_refresh-cw", iconSize=16, tooltip="Refresh the UI"
-        )
-        rbf_cbox.setFixedHeight(self.genericWidgetHight)
-        rbf_refreshButton.setMaximumWidth(80)
-        rbf_refreshButton.setFixedHeight(self.genericWidgetHight - 1)
-        setRBFLayout.addWidget(rbfLabel)
-        setRBFLayout.addWidget(rbf_cbox, 1)
-        setRBFLayout.addWidget(rbf_refreshButton)
-        return setRBFLayout, rbf_cbox, rbf_refreshButton
-
-    def createDriverAttributeWidget(self):
-        """widget where the user inputs information for the setups
-
-        Returns:
-            list: [of widgets]
-        """
-        driverControlVLayout = QtWidgets.QVBoxLayout()
-        driverControlHLayout = QtWidgets.QHBoxLayout()
-
-        # driverMainLayout.setStyleSheet("QVBoxLayout { background-color: #404040;")
-        driverControlHLayout.setSpacing(3)
-        #  --------------------------------------------------------------------
-        (controlLayout,
-         controlLineEdit,
-         setControlButton) = self.selectNodeWidget("Control", buttonLabel="Set")
-        controlLineEdit.setToolTip("The node driving the setup. (Click me!)")
-        #  --------------------------------------------------------------------
-        (driverLayout,
-         driverLineEdit,
-         driverSelectButton) = self.selectNodeWidget("Driver", buttonLabel="Set")
-        driverLineEdit.setToolTip("The node driving the setup. (Click me!)")
-        #  --------------------------------------------------------------------
-        allButton = self.createCustomButton("", (20, 52), tooltip="", icon="mgear_rewind", iconSize=15)
-
-        (attributeLayout, attributeListWidget) = self.labelListWidget(
-            label="Select Driver Attributes:", attrListType="driver", horizontal=False)
-
-        attributeListWidget.setToolTip("List of attributes driving setup.")
-        selType = QtWidgets.QAbstractItemView.ExtendedSelection
-        attributeListWidget.setSelectionMode(selType)
-        attributeListWidget.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
-        #  --------------------------------------------------------------------
-        driverControlVLayout.addLayout(controlLayout, 0)
-        driverControlVLayout.addLayout(driverLayout, 0)
-        driverControlHLayout.addLayout(driverControlVLayout, 0)
-        driverControlHLayout.addWidget(allButton, 0)
-        return [controlLineEdit,
-                setControlButton,
-                driverLineEdit,
-                driverSelectButton,
-                allButton,
-                attributeListWidget,
-                attributeLayout,
-                driverControlHLayout]
-
-    def createDrivenAttributeWidget(self):
-        """the widget that displays the driven information
-
-        Returns:
-            list: [of widgets]
-        """
-        drivenWidget = QtWidgets.QWidget()
-        drivenMainLayout = QtWidgets.QVBoxLayout()
-        drivenMainLayout.setContentsMargins(0, 10, 0, 2)
-        drivenMainLayout.setSpacing(9)
-        drivenSetLayout = QtWidgets.QVBoxLayout()
-        drivenMainLayout.addLayout(drivenSetLayout)
-        drivenWidget.setLayout(drivenMainLayout)
-        #  --------------------------------------------------------------------
-        (drivenLayout,
-         drivenLineEdit,
-         drivenSelectButton) = self.selectNodeWidget("Driven", buttonLabel="Set")
-        drivenTip = "The node being driven by setup. (Click me!)"
-        drivenLineEdit.setToolTip(drivenTip)
-
-        addDrivenButton = self.createCustomButton("", (20, 25), icon="mgear_plus", iconSize=16, tooltip="")
-        addDrivenButton.setToolTip("Add a new driven to the current rbf node")
-        #  --------------------------------------------------------------------
-        (attributeLayout,
-         attributeListWidget) = self.labelListWidget(label="Select Driven Attributes:",
-                                                     attrListType="driven",
-                                                     horizontal=False)
-        attributeListWidget.setToolTip("Attributes being driven by setup.")
-        attributeLayout.setSpacing(1)
-        selType = QtWidgets.QAbstractItemView.ExtendedSelection
-        attributeListWidget.setSelectionMode(selType)
-        attributeListWidget.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
-        #  --------------------------------------------------------------------
-        drivenSetLayout.addLayout(drivenLayout, 0)
-        drivenSetLayout.addLayout(attributeLayout, 0)
-        drivenLayout.addWidget(addDrivenButton)
-        drivenSetLayout.addLayout(drivenLayout, 0)
-        drivenSetLayout.addLayout(attributeLayout, 0)
-        drivenMainLayout.addLayout(drivenSetLayout)
-        drivenWidget.setLayout(drivenMainLayout)
-        return [drivenLineEdit,
-                drivenSelectButton,
-                addDrivenButton,
-                attributeListWidget,
-                drivenWidget,
-                drivenMainLayout]
-
-    def createDrivenWidget(self):
-        """the widget that displays the driven information
-
-        Returns:
-            list: [of widgets]
-        """
-        drivenWidget = QtWidgets.QWidget()
-        drivenMainLayout = QtWidgets.QHBoxLayout()
-        drivenMainLayout.setContentsMargins(0, 0, 0, 0)
-        drivenMainLayout.setSpacing(9)
-        drivenWidget.setLayout(drivenMainLayout)
-
-        tableWidget = self.createTableWidget()
-        drivenMainLayout.addWidget(tableWidget, 1)
-        return drivenWidget, tableWidget
-
-    def createTableWidget(self):
-        """create table widget used to display poses, set tooltips and colum
-
-        Returns:
-            QTableWidget: QTableWidget
-        """
-        stylesheet = """
-        QTableWidget QHeaderView::section {
-            background-color: #3a3b3b;
-            padding: 2px;
-            text-align: center;
-        }
-        QTableCornerButton::section {
-            background-color: #3a3b3b;
-            border: none; 
-        }
-        """
-        tableWidget = QtWidgets.QTableWidget()
-        tableWidget.insertColumn(0)
-        tableWidget.insertRow(0)
-        tableWidget.setHorizontalHeaderLabels(["Pose Value"])
-        tableWidget.setVerticalHeaderLabels(["Pose #0"])
-        # tableWidget.setStyleSheet(stylesheet)
-        tableTip = "Live connections to the RBF Node in your setup."
-        tableTip = tableTip + "\nSelect the desired Pose # to recall pose."
-        tableWidget.setToolTip(tableTip)
-        return tableWidget
-
-    def createTabWidget(self):
-        """Tab widget to add driven widgets too. Custom TabBar so the tab is
-        easier to select
-
-        Returns:
-            QTabWidget:
-        """
-        tabLayout = QtWidgets.QTabWidget()
-        tabLayout.setContentsMargins(0, 0, 0, 0)
-        tabBar = TabBar()
-        tabLayout.setTabBar(tabBar)
-        tabBar.setTabsClosable(True)
-        return tabLayout
-
-    def createOptionsButtonsWidget(self):
-        """add, edit, delete buttons for modifying rbf setups.
-
-        Returns:
-            list: [QPushButtons]
-        """
-        optionsLayout = QtWidgets.QHBoxLayout()
-        optionsLayout.setSpacing(5)
-        addTip = "After positioning all controls in the setup, add new pose."
-        addTip = addTip + "\nEnsure the driver node has a unique position."
-        addPoseButton = self.createCustomButton("Add Pose", size=(80, 26), tooltip=addTip)
-        EditPoseButton = self.createCustomButton("Update Pose", size=(80, 26))
-        EditPoseButton.setToolTip("Recall pose, adjust controls and Edit.")
-        EditPoseValuesButton = self.createCustomButton("Update Pose Values", size=(80, 26))
-        EditPoseValuesButton.setToolTip("Set pose based on values in table")
-        deletePoseButton = self.createCustomButton("Delete Pose", size=(80, 26))
-        deletePoseButton.setToolTip("Recall pose, then Delete")
-        optionsLayout.addWidget(addPoseButton)
-        optionsLayout.addWidget(EditPoseButton)
-        optionsLayout.addWidget(EditPoseValuesButton)
-        optionsLayout.addWidget(deletePoseButton)
-        return (optionsLayout,
-                addPoseButton,
-                EditPoseButton,
-                EditPoseValuesButton,
-                deletePoseButton)
-
-    def createDarkContainerWidget(self):
-        darkContainer = QtWidgets.QWidget()
-        driverMainLayout = QtWidgets.QVBoxLayout()
-        driverMainLayout.setContentsMargins(10, 10, 10, 10)  # Adjust margins as needed
-        driverMainLayout.setSpacing(5)  # Adjust spacing between widgets
-
-        # Setting the dark color (Example: dark gray)
-        # darkContainer.setStyleSheet("background-color: #323232;")
-
-        # Driver section
-        (self.controlLineEdit,
-         self.setControlButton,
-         self.driverLineEdit,
-         self.setDriverButton,
-         self.allButton,
-         self.driverAttributesWidget,
-         self.driverAttributesLayout,
-         driverControlLayout) = self.createDriverAttributeWidget()
-
-        # Driven section
-        (self.drivenLineEdit,
-         self.setDrivenButton,
-         self.addDrivenButton,
-         self.drivenAttributesWidget,
-         self.drivenWidget,
-         self.drivenMainLayout) = self.createDrivenAttributeWidget()
-
-        self.addRbfButton = self.createCustomButton("New RBF")
-        self.addRbfButton.setToolTip("Select node to be driven by setup.")
-        stylesheet = (
-            "QPushButton {background-color: #179e83; border-radius: 4px;}"
-            "QPushButton:hover:!pressed { background-color: #2ea88f;}"
-        )
-        self.addRbfButton.setStyleSheet(stylesheet)
-
-        # Setting up the main layout for driver and driven sections
-        driverMainLayout = QtWidgets.QVBoxLayout()
-        driverMainLayout.addLayout(driverControlLayout)
-        driverMainLayout.addLayout(self.driverAttributesLayout)
-        driverMainLayout.addWidget(self.drivenWidget)
-        driverMainLayout.addWidget(self.addRbfButton)
-        darkContainer.setLayout(driverMainLayout)
-
-        return darkContainer
-
-    def createDriverDrivenTableWidget(self):
-        tableContainer = QtWidgets.QWidget()
-
-        # Setting up the main layout for driver and driven sections
-        driverDrivenTableLayout = QtWidgets.QVBoxLayout()
-        self.driverPoseTableWidget = self.createTableWidget()
-        self.rbfTabWidget = self.createTabWidget()
-
-        # Options buttons section
-        (optionsLayout,
-         self.addPoseButton,
-         self.editPoseButton,
-         self.editPoseValuesButton,
-         self.deletePoseButton) = self.createOptionsButtonsWidget()
-        self.addPoseButton.setEnabled(False)
-        self.editPoseButton.setEnabled(False)
-        self.editPoseValuesButton.setEnabled(False)
-        self.deletePoseButton.setEnabled(False)
-
-        driverDrivenTableLayout.addWidget(self.driverPoseTableWidget, 1)
-        driverDrivenTableLayout.addWidget(self.rbfTabWidget, 1)
-        driverDrivenTableLayout.addLayout(optionsLayout)
-        tableContainer.setLayout(driverDrivenTableLayout)
-
-        return tableContainer
-
-
-class RBFTables(RBFWidget):
+class RBFTables(widget.RBFWidget):
 
     def __init__(self):
         pass
 
 
-class RBFManagerUI(RBFWidget):
+class RBFMenuFunction:
 
+    def __init__(self, rbfInstance):
+        self.rbfInstance = rbfInstance
+
+    def deleteSetup(self, setupName=None, allSetup=False):
+        """Delete all the nodes within a setup.
+
+        Args:
+            setupName (None, optional): Description
+        """
+        decision = widget.promptAcceptance(self.rbfInstance,
+                                           "Delete current Setup?",
+                                           "This will delete all RBF nodes in setup.")
+        if decision in [QtWidgets.QMessageBox.Discard,
+                        QtWidgets.QMessageBox.Cancel]:
+            return
+
+        nodesToDelete = None
+        if setupName is None:
+            if not allSetup:
+                setupName, _ = self.rbfInstance.getSelectedSetup()
+                nodesToDelete = self.rbfInstance.allSetupsInfo.get(setupName, [])
+            else:
+                nodesToDelete = list(self.rbfInstance.allSetupsInfo.values())
+                nodesToDelete = [item for sublist in nodesToDelete for item in sublist]
+
+        for rbfNode in nodesToDelete:
+            drivenNode = rbfNode.getDrivenNode()
+            rbfNode.deleteRBFToggleAttr()
+            if drivenNode:
+                rbf_node.removeDrivenGroup(drivenNode[0])
+            mc.delete(rbfNode.transformNode)
+        self.rbfInstance.refresh()
+
+    def importNodes(self):
+        """import a setup(s) from file select by user
+
+        Returns:
+            n/a: nada
+        """
+        filePath = rbf_io.fileDialog(mode=1)
+        if filePath is None:
+            return
+        rbf_io.importRBFs(filePath)
+        mc.select(cl=True)
+        self.rbfInstance.refresh()
+        print("RBF setups imported: {}".format(filePath))
+
+    def exportNodes(self, allSetups=True):
+        """export all nodes or nodes from current setup
+
+        Args:
+            allSetups (bool, optional): If all or setup
+
+        Returns:
+            n/a: nada
+        """
+        # TODO WHEN NEW RBF NODE TYPES ARE ADDED, THIS WILL NEED TO BE RETOOLED
+        nodesToExport = []
+        if allSetups:
+            [nodesToExport.extend(v) for k, v,
+             in self.rbfInstance.allSetupsInfo.items()]
+        else:
+            nodesToExport = self.rbfInstance.currentRBFSetupNodes
+
+        nodesToExport = [n.name for n in nodesToExport]
+        filePath = rbf_io.fileDialog(mode=0)
+        if filePath is None:
+            return
+        rbf_io.exportRBFs(nodesToExport, filePath)
+
+
+class RBFManagerUI(widget.RBFWidget):
     """A manager for creating, mirroring, importing/exporting poses created
     for RBF type nodes.
 
@@ -818,6 +328,8 @@ class RBFManagerUI(RBFWidget):
         self.drivenWidget = []
         self.driverAutoAttr = []
         self.drivenAutoAttr = []
+
+        self.menuFunc = RBFMenuFunction(self)
 
         self.setMenuBar(self.createMenuBar(hideMenuBar=hideMenuBar))
         self.setCentralWidget(self.createCentralWidget())
@@ -906,33 +418,6 @@ class RBFManagerUI(RBFWidget):
             drivenNodes.extend(rbfNode.getDrivenNode)
         return drivenNodes
 
-    def __deleteSetup(self):
-        decision = promptAcceptance(self,
-                                    "Delete current Setup?",
-                                    "This will delete all RBF nodes in setup.")
-        if decision in [QtWidgets.QMessageBox.Discard,
-                        QtWidgets.QMessageBox.Cancel]:
-            return
-        self.deleteSetup()
-
-    def deleteSetup(self, setupName=None):
-        """Delete all the nodes within a setup.
-
-        Args:
-            setupName (None, optional): Description
-        """
-        setupType = None
-        if setupName is None:
-            setupName, setupType = self.getSelectedSetup()
-        nodesToDelete = self.allSetupsInfo.get(setupName, [])
-        for rbfNode in nodesToDelete:
-            drivenNode = rbfNode.getDrivenNode()
-            rbfNode.deleteRBFToggleAttr()
-            if drivenNode:
-                rbf_node.removeDrivenGroup(drivenNode[0])
-            mc.delete(rbfNode.transformNode)
-        self.refresh()
-
     def removeRBFFromSetup(self, drivenWidgetIndex):
         """remove RBF tab from setup. Delete driven group, attrs and clean up
 
@@ -943,9 +428,9 @@ class RBFManagerUI(RBFWidget):
         Returns:
             n/a: n/a
         """
-        decision = promptAcceptance(self,
-                                    "Are you sure you want to remove node?",
-                                    "This will delete the RBF & driven node.")
+        decision = widget.promptAcceptance(self,
+                                           "Are you sure you want to remove node?",
+                                           "This will delete the RBF & driven node.")
         if decision in [QtWidgets.QMessageBox.Discard,
                         QtWidgets.QMessageBox.Cancel]:
             return
@@ -991,7 +476,7 @@ class RBFManagerUI(RBFWidget):
         if mc.objExists(drivenNode_name):
             if existing_rbf_setup(drivenNode_name):
                 msg = "Node is already driven by an RBF Setup."
-                genericWarning(self, msg)
+                widget.genericWarning(self, msg)
                 return
 
         setupName, rbfType = self.getSelectedSetup()
@@ -1066,7 +551,7 @@ class RBFManagerUI(RBFWidget):
 
         # Check if the driven node is the same as the driver node
         if drivenNode == driverNode:
-            genericWarning(self, "Select Node to be driven!")
+            widget.genericWarning(self, "Select Node to be driven!")
             return result
 
         # Update the result dictionary with the fetched values
@@ -1123,7 +608,7 @@ class RBFManagerUI(RBFWidget):
         # TODO if one is allow syncing of nodes of different lengths
         # it should be done here
         if drivenRow != driverRow or drivenRow == -1:
-            genericWarning(self, "Select Pose # to be deleted.")
+            widget.genericWarning(self, "Select Pose # to be deleted.")
             return
 
         for rbfNode in self.currentRBFSetupNodes:
@@ -1144,7 +629,7 @@ class RBFManagerUI(RBFWidget):
         drivenTableWidget = getattr(drivenWidget, "tableWidget")
         drivenRow = drivenTableWidget.currentRow()
         if drivenRow != driverRow or drivenRow == -1:
-            genericWarning(self, "Select Pose # to be Edited.")
+            widget.genericWarning(self, "Select Pose # to be Edited.")
             return
         driverNode = rbfNodes[0].getDriverNode()[0]
         driverAttrs = rbfNodes[0].getDriverNodeAttributes()
@@ -1165,12 +650,12 @@ class RBFManagerUI(RBFWidget):
         rbfNodes = self.currentRBFSetupNodes
         if not rbfNodes:
             return
-        driverRow = self.driverPoseTableWidget.currentRow() # A number
+        driverRow = self.driverPoseTableWidget.currentRow()  # A number
         drivenWidget = self.rbfTabWidget.currentWidget()
         drivenTableWidget = getattr(drivenWidget, "tableWidget")
         drivenRow = drivenTableWidget.currentRow()
         if drivenRow != driverRow or drivenRow == -1:
-            genericWarning(self, "Select Pose # to be Edited.")
+            widget.genericWarning(self, "Select Pose # to be Edited.")
             return
         driverNode = rbfNodes[0].getDriverNode()[0]
         driverAttrs = rbfNodes[0].getDriverNodeAttributes()
@@ -1432,7 +917,7 @@ class RBFManagerUI(RBFWidget):
                     rbfAttrPlug = "{}.poses[{}].poseInput[{}]".format(rbfNode[0], rowIndex, columnIndex)
 
                     # Create a control widget for this pose input attribute
-                    getControlAttrWidget(rbfAttrPlug, label="")
+                    widget.getControlAttrWidget(rbfAttrPlug, label="")
 
     def setDriverTable(self, rbfNode, weightInfo):
         """Set the driverTable widget with the information from the weightInfo
@@ -1470,7 +955,15 @@ class RBFManagerUI(RBFWidget):
         for rowIndex, poseInput in enumerate(poses["poseInput"]):
             for columnIndex, _ in enumerate(poseInput):
                 rbfAttrPlug = "{}.poses[{}].poseInput[{}]".format(rbfNode, rowIndex, columnIndex)
-                attrEdit, mAttrField = getControlAttrWidget(rbfAttrPlug, label="")
+                attrEdit, mAttrField = widget.getControlAttrWidget(rbfAttrPlug, label="")
+
+                # Get the current width of the column after resizing to contents
+                self.driverPoseTableWidget.resizeColumnToContents(columnIndex)
+                currentWidth = self.driverPoseTableWidget.columnWidth(columnIndex)
+
+                # Add some extra space (e.g., 10 pixels)
+                extraSpace = 20
+                self.driverPoseTableWidget.setColumnWidth(columnIndex, currentWidth + extraSpace)
 
                 self.driverPoseTableWidget.setCellWidget(rowIndex, columnIndex, attrEdit)
                 attrEdit.returnPressed.connect(
@@ -1582,8 +1075,16 @@ class RBFManagerUI(RBFWidget):
         for rowIndex, poseInput in enumerate(poses["poseValue"]):
             for columnIndex, pValue in enumerate(poseInput):
                 rbfAttrPlug = "{}.poses[{}].poseValue[{}]".format(rbfNode, rowIndex, columnIndex)
-                attrEdit, mAttrFeild = getControlAttrWidget(rbfAttrPlug, label="")
+                attrEdit, mAttrFeild = widget.getControlAttrWidget(rbfAttrPlug, label="")
                 drivenWidget.tableWidget.setCellWidget(rowIndex, columnIndex, attrEdit)
+
+                # Get the current width of the column after resizing to contents
+                drivenWidget.tableWidget.resizeColumnToContents(columnIndex)
+                currentWidth = drivenWidget.tableWidget.columnWidth(columnIndex)
+
+                # Add some extra space (e.g., 10 pixels)
+                extraSpace = 20
+                drivenWidget.tableWidget.setColumnWidth(columnIndex, currentWidth + extraSpace)
 
                 # Adding QTableWidgetItem for the cell and setting the value
                 tableItem = QtWidgets.QTableWidgetItem()
@@ -1636,7 +1137,7 @@ class RBFManagerUI(RBFWidget):
             drivenSelection=True,
             currentRBFSetupNodes=False,
             clearDrivenTab=False
-            )
+        )
 
         self.setDrivenButton.blockSignals(False)
         self.drivenAttributesWidget.setEnabled(True)
@@ -1908,12 +1409,30 @@ class RBFManagerUI(RBFWidget):
         elif self.currentRBFSetupNodes:
             textA = "Do you want to change the Control for setup?"
             textB = "This Control that will be used for recalling poses."
-            decision = promptAcceptance(self, textA, textB)
+            decision = widget.promptAcceptance(self, textA, textB)
             if decision in [QtWidgets.QMessageBox.Discard,
                             QtWidgets.QMessageBox.Cancel]:
                 return
             controlName = self.setNodeToField(lineEditWidget)
             self.setDriverControlOnSetup(controlName)
+
+    def setDrivenInfo(self, tabIndex):
+        self.refresh(rbfSelection=False,
+                     driverSelection=False,
+                     drivenSelection=True,
+                     currentRBFSetupNodes=False,
+                     clearDrivenTab=False)
+
+        tabWidget = self.rbfTabWidget.widget(tabIndex)
+        rbfNode = getattr(tabWidget, "rbfNode")
+        weightInfo = rbfNode.getNodeInfo()
+
+        # Set Driven Node and Attributes
+        drivenNode = weightInfo.get("drivenNode", [None])[0]
+        self.drivenLineEdit.setText(drivenNode or "")
+        self.setAttributeDisplay(self.drivenAttributesWidget, drivenNode or "", weightInfo["drivenAttrs"])
+
+        self.addPoseButton.setEnabled(True)
 
     @staticmethod
     def getRBFNodesInfo(rbfNodes):
@@ -1930,47 +1449,6 @@ class RBFManagerUI(RBFWidget):
         for rbf in rbfNodes:
             weightNodeInfo_dict[rbf.name] = rbf.getNodeInfo()
         return weightNodeInfo_dict
-
-    def importNodes(self):
-        """import a setup(s) from file select by user
-
-        Returns:
-            n/a: nada
-        """
-        # sceneFilePath = mc.file(sn=True, q=True)
-        # startDir = os.path.dirname(sceneFilePath)
-        filePath = rbf_io.fileDialog(mode=1)
-        if filePath is None:
-            return
-        rbf_io.importRBFs(filePath)
-        mc.select(cl=True)
-        self.refresh()
-        print("RBF setups imported: {}".format(filePath))
-
-    def exportNodes(self, allSetups=True):
-        """export all nodes or nodes from current setup
-
-        Args:
-            allSetups (bool, optional): If all or setup
-
-        Returns:
-            n/a: nada
-        """
-        # TODO WHEN NEW RBF NODE TYPES ARE ADDED, THIS WILL NEED TO BE RETOOLED
-        nodesToExport = []
-        if allSetups:
-            [nodesToExport.extend(v) for k, v,
-             in self.allSetupsInfo.items()]
-        else:
-            nodesToExport = self.currentRBFSetupNodes
-
-        nodesToExport = [n.name for n in nodesToExport]
-        # sceneFilePath = mc.file(sn=True, q=True)
-        # startDir = os.path.dirname(sceneFilePath)
-        filePath = rbf_io.fileDialog(mode=0)
-        if filePath is None:
-            return
-        rbf_io.exportRBFs(nodesToExport, filePath)
 
     @staticmethod
     def gatherMirroredInfo(rbfNodes):
@@ -2002,7 +1480,7 @@ class RBFManagerUI(RBFWidget):
             # setupName -------------------------------------------------------
             mrSetupName = mString.convertRLName(weightInfo["setupName"])
             if mrSetupName == weightInfo["setupName"]:
-                mrSetupName = "{}{}".format(mrSetupName, MIRROR_SUFFIX)
+                mrSetupName = "{}{}".format(mrSetupName, widget.MIRROR_SUFFIX)
             weightInfo["setupName"] = mrSetupName
             # transformNode ---------------------------------------------------
             tmp = weightInfo["transformNode"]["name"]
@@ -2077,7 +1555,7 @@ class RBFManagerUI(RBFWidget):
             mrData = []
             for srcNode, dstValues in setupTargetInfo_dict.items():
                 mrData.extend(anim_utils.calculateMirrorDataRBF(srcNode,
-                                                             dstValues[0]))
+                                                                dstValues[0]))
             for entry in mrData:
                 anim_utils.applyMirror(nameSpace, entry)
             poseInputs = rbf_node.getMultipleAttrs(mrDriverNode, mrDriverAttrs)
@@ -2166,6 +1644,7 @@ class RBFManagerUI(RBFWidget):
         # Driver Line Edit and Control Line Edit
         self.driverLineEdit.clicked.connect(selectNode)
         self.controlLineEdit.clicked.connect(selectNode)
+        self.drivenLineEdit.clicked.connect(selectNode)
         self.driverLineEdit.textChanged.connect(partial(self.updateAttributeDisplay,
                                                         self.driverAttributesWidget))
         self.drivenLineEdit.textChanged.connect(partial(self.updateAttributeDisplay,
@@ -2210,6 +1689,7 @@ class RBFManagerUI(RBFWidget):
         tabBar = self.rbfTabWidget.tabBar()
         tabBar.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
         tabBar.customContextMenuRequested.connect(self.tabConextMenu)
+        tabBar.tabBarClicked.connect(self.setDrivenInfo)
         tabBar.tabCloseRequested.connect(self.removeRBFFromSetup)
 
     # main assebly ------------------------------------------------------------
@@ -2236,15 +1716,15 @@ class RBFManagerUI(RBFWidget):
         allTableWidget = self.createDriverDrivenTableWidget()
 
         centralWidgetLayout.addLayout(rbfLayout)
-        centralWidgetLayout.addWidget(HLine())
+        centralWidgetLayout.addWidget(widget.HLine())
         splitter.addWidget(driverDrivenWidget)
         splitter.addWidget(allTableWidget)
         centralWidgetLayout.addWidget(splitter)
 
         # Assuming a ratio of 2:1 for settingWidth to tableWidth
         totalWidth = splitter.width()
-        attributeWidth = (1/3) * totalWidth
-        tableWidth = (2/3) * totalWidth
+        attributeWidth = (1 / 3) * totalWidth
+        tableWidth = (2 / 3) * totalWidth
         splitter.setSizes([int(attributeWidth), int(tableWidth)])
         return centralWidget
 
@@ -2262,12 +1742,12 @@ class RBFManagerUI(RBFWidget):
         file = mainMenuBar.addMenu("File")
         menu1 = file.addAction("Re-evaluate Nodes", self.reevalluateAllNodes)
         menu1.setToolTip("Force all RBF nodes to re-revaluate.")
-        file.addAction("Export All", self.exportNodes)
-        file.addAction("Export current setup", partial(self.exportNodes,
-                                                       allSetups=False))
-        file.addAction("Import RBFs", partial(self.importNodes))
         file.addSeparator()
-        file.addAction("Delete Current Setup", self.__deleteSetup)
+        file.addAction("Import RBFs", partial(self.menuFunc.importNodes))
+        file.addAction("Export RBFs", self.menuFunc.exportNodes)
+        file.addSeparator()
+        file.addAction("Delete Current Setup", partial(self.menuFunc.deleteSetup, allSetup=False))
+        file.addAction("Delete All Setup", partial(self.menuFunc.deleteSetup, allSetup=True))
         # mirror --------------------------------------------------------------
         mirrorMenu = mainMenuBar.addMenu("Mirror")
         mirrorMenu1 = mirrorMenu.addAction("Mirror Setup", self.mirrorSetup)
@@ -2317,4 +1797,3 @@ class RBFManagerUI(RBFWidget):
             if event.buttons() == QtCore.Qt.NoButton:
                 pos = event.pos()
                 self.mousePosition.emit(pos.x(), pos.y())
-
